extracted constants to GenesisCityData static class
